Rewiring of the UI (over navigation) and Survey Reducers.
