+++ src/common/actionTypes/uiActionTypes.js
@@ -1,0 +1,4 @@
+export const INVITE_USER = 'INVITE_USER';
+export const SUBNAVIGATE = 'SUBNAVIGATE';
+export const UPDATE_USER_SEARCH_GROUP = 'UPDATE_USER_SEARCH_GROUP';
+export const UPDATE_USER_SEARCH_QUERY = 'UPDATE_USER_SEARCH_QUERY';
