+++ src/common/actions/index.js
@@ -1,0 +1,1 @@
+export * from './uiActions';
