+++ src/common/actions/uiActions.js
@@ -1,0 +1,30 @@
+import * as actionTypes from '../actionTypes/uiActionTypes';
+
+export function subnavigate(id) {
+    return {
+        type: actionTypes.SUBNAVIGATE,
+        id,
+    };
+}
+
+export function inviteUser(user, projectId) {
+    return {
+        type: actionTypes.INVITE_USER,
+        user,
+        projectId,
+    };
+}
+
+export function updateUserSearchGroup(group) {
+    return {
+        type: actionTypes.UPDATE_USER_SEARCH_GROUP,
+        group
+    };
+}
+
+export function updateUserSearchQuery(query) {
+    return {
+        type: actionTypes.UPDATE_USER_SEARCH_QUERY,
+        query
+    };
+}
