+++ src/common/reducers/index.js
@@ -2,3 +2,4 @@ export { NavReducer } from './navReducer';
 export { SettingsReducer } from './settingsReducer';
 export { UserReducer } from './userReducer';
 export { SurveysReducer } from './surveysReducer';
+export { UIReducer } from './uiReducer';
