+++ src/common/reducers/uiReducer.js
@@ -1,4 +1,6 @@
 import en from '../../i18n/en';
+import update from 'immutability-helper';
+import * as type from '../actionTypes/uiActionTypes';
 
 const initialState = {
     subnav: 'workflow',
@@ -6,18 +8,20 @@ const initialState = {
         query: '',
         groups: {},
     },
-    // Look into Redux Forms
-}
+};
 
-export const uiReducer = (state = initialState, action) => {
-    case type.SUBNAVIGATE:
-        return update(state, { ui: { subnav: { $set: action.id } } });
-    case type.UPDATE_USER_SEARCH_GROUP:
-        return(update(state, { ui: { userSidebarSearch: {
-            group: { $set: action.group },
-        } } } ) );
-    case type.UPDATE_USER_SEARCH_QUERY:
-        return update(state, { ui: { userSidebarSearch: {
-            query: { $set: action.query } } } } );
-    return state;
+export const UIReducer = (state = initialState, action) => {
+    switch (action.type) {
+        case type.SUBNAVIGATE:
+            return update(state, { subnav: { $set: action.id } });
+        case type.UPDATE_USER_SEARCH_GROUP:
+            return update(state, { userSidebarSearch: {
+                group: { $set: action.group },
+            } } );
+        case type.UPDATE_USER_SEARCH_QUERY:
+            return update(state, { userSidebarSearch: {
+                query: { $set: action.query } } } );
+        default:
+            return state;
+    }
 };
