+++ src/views/ProjectManagement/actionTypes.js
@@ -1,6 +1,5 @@
 // Summary reducer calls.
 export const EDIT_SUMMARY_DESCRIPTION = 'EDIT_SUMMARY_DESCRIPTION';
-export const SUBNAVIGATE = 'SUBNAVIGATE';
 export const ASSIGN_TASK = 'ASSIGN_TASK';
 export const ADD_SUBJECT = 'ADD_SUBJECT';
 export const DELETE_SUBJECT = 'DELETE_SUBJECT';
@@ -9,8 +8,3 @@ export const UPDATE_TASK = 'UPDATE_TASK';
 export const TOGGLE_FILTER = 'TOGGLE_FILTER';
 export const SET_PROJECT_STATUS = 'SET_PROJECT_STATUS';
 export const SET_SURVEY_STATUS = 'SET_SURVEY_STATUS';
-
-// User Sidebar:
-export const INVITE_USER = 'INVITE_USER';
-export const UPDATE_USER_SEARCH_GROUP = 'UPDATE_USER_SEARCH_GROUP';
-export const UPDATE_USER_SEARCH_QUERY = 'UPDATE_USER_SEARCH_QUERY';
